API de ventas 90%, productos 100%, sucursales 100%, cajas 50%

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -20,7 +20,7 @@ class CreateUsersTable extends Migration
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->string('avatar')->nullable()->default('../images/user.svg');
-            $table->integer('status')->nullable()->default(0);
+            $table->smallInteger('status')->nullable()->default(0);
             $table->string('firebase_token')->nullable();
             $table->rememberToken();
             $table->timestamps();
